move parsing logic
