+++ extension/src/plots/model/collect.test.ts
@@ -9,8 +9,7 @@ import {
   collectTemplates,
   collectMetricOrder,
   collectWorkspaceRunningCheckpoint,
-  collectWorkspaceRaceConditionData,
-  collectMultiSourceData
+  collectWorkspaceRaceConditionData
 } from './collect'
 import plotsDiffFixture from '../../test/fixtures/plotsDiff/output'
 import expShowFixture from '../../test/fixtures/expShow/output'
@@ -357,175 +356,5 @@ describe('collectWorkspaceRaceConditionData', () => {
         values.map(value => omit(value, 'rev'))
       )
     )
-  })
-})
-
-describe('collectMultiSourceData', () => {
-  it('should return an empty object given no multi source data', () => {
-    const multiSourceKeys = collectMultiSourceData({
-      path: [{ field: 'x', filename: 'path' }]
-    })
-    expect(multiSourceKeys).toStrictEqual({})
-  })
-
-  it('should return an object containing a filename strokeDash given data with varying filenames', () => {
-    const otherPath = join('other', 'path')
-    const multiSourceKeys = collectMultiSourceData({
-      combined: [
-        { field: 'x', filename: 'path' },
-        { field: 'x', filename: otherPath }
-      ]
-    })
-    expect(multiSourceKeys).toStrictEqual({
-      combined: {
-        strokeDash: {
-          field: 'filename',
-          scale: {
-            domain: [otherPath, 'path'],
-            range: [
-              [1, 0],
-              [8, 8]
-            ]
-          }
-        }
-      }
-    })
-  })
-
-  it('should return an object containing a field strokeDash given data with varying fields', () => {
-    const multiSourceKeys = collectMultiSourceData({
-      path: [
-        { field: 'x', filename: 'path' },
-        { field: 'z', filename: 'path' }
-      ]
-    })
-    expect(multiSourceKeys).toStrictEqual({
-      path: {
-        strokeDash: {
-          field: 'field',
-          scale: {
-            domain: ['x', 'z'],
-            range: [
-              [1, 0],
-              [8, 8]
-            ]
-          }
-        }
-      }
-    })
-  })
-
-  it('should return an object containing a filename and field strokeDash given data with varying filename and fields', () => {
-    const otherPath = join('other', 'path')
-    const multiSourceKeys = collectMultiSourceData({
-      combined: [
-        { field: 'x', filename: 'path' },
-        { field: 'z', filename: otherPath }
-      ]
-    })
-    expect(multiSourceKeys).toStrictEqual({
-      combined: {
-        strokeDash: {
-          field: 'filename::field',
-          scale: {
-            domain: [`${otherPath}::z`, 'path::x'],
-            range: [
-              [1, 0],
-              [8, 8]
-            ]
-          }
-        }
-      }
-    })
-  })
-
-  it('should return an object containing a filename strokeDash and field shape given data with differently varying filename and field', () => {
-    const multiSourceKeys = collectMultiSourceData({
-      combined: [
-        { field: 'x', filename: 'path' },
-        { field: 'z', filename: 'path' },
-        { field: 'z', filename: join('other', 'path') }
-      ]
-    })
-    expect(multiSourceKeys).toStrictEqual({
-      combined: {
-        shape: {
-          field: 'field',
-          scale: { domain: ['x', 'z'], range: ['square', 'circle'] }
-        },
-        strokeDash: {
-          field: 'filename',
-          scale: {
-            domain: [join('other', 'path'), 'path'],
-            range: [
-              [1, 0],
-              [8, 8]
-            ]
-          }
-        }
-      }
-    })
-  })
-
-  it('should return an object containing a filename strokeDash and field shape given data with varying filename and similar field', () => {
-    const multiSourceKeys = collectMultiSourceData({
-      combined: [
-        { field: 'x', filename: join('first', 'path') },
-        { field: 'z', filename: join('second', 'path') },
-        { field: 'z', filename: join('third', 'path') }
-      ]
-    })
-    expect(multiSourceKeys).toStrictEqual({
-      combined: {
-        shape: {
-          field: 'field',
-          scale: { domain: ['x', 'z'], range: ['square', 'circle'] }
-        },
-        strokeDash: {
-          field: 'filename',
-          scale: {
-            domain: [
-              join('first', 'path'),
-              join('second', 'path'),
-              join('third', 'path')
-            ],
-            range: [
-              [1, 0],
-              [8, 8],
-              [8, 4]
-            ]
-          }
-        }
-      }
-    })
-  })
-
-  it('should return an object containing a filename strokeDash and field shape given data with varying filename and similar field 2', () => {
-    const multiSourceKeys = collectMultiSourceData({
-      combined: [
-        { field: 'x', filename: join('first', 'path') },
-        { field: 'z', filename: join('second', 'path') },
-        { field: 'q', filename: join('third', 'path') }
-      ]
-    })
-    expect(multiSourceKeys).toStrictEqual({
-      combined: {
-        strokeDash: {
-          field: 'filename::field',
-          scale: {
-            domain: [
-              `${join('first', 'path')}::x`,
-              `${join('second', 'path')}::z`,
-              `${join('third', 'path')}::q`
-            ],
-            range: [
-              [1, 0],
-              [8, 8],
-              [8, 4]
-            ]
-          }
-        }
-      }
-    })
   })
 })
